refactor: move RiskDisclosure to common
