+++ packages/bzx-common/src/shared-components/RiskDisclosure/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RiskDisclosure'
