+++ packages/bzx-common/src/styles/index.scss
@@ -1,0 +1,1 @@
+@import './../ui-framework/styles/variables.scss';
