+++ packages/bzx-common/src/ui-framework/styles/variables.scss
@@ -1,0 +1,30 @@
+$page-max-width: 76rem;
+$page-padding: 0 2rem;
+$mobile-version-max-width: 59.9375rem;
+$mobile-version-width: 600px;
+$mobile-version-width-sm: 1024px;
+
+$container-desktop: 1380px;
+$container-laptop: 1110px;
+$container-mobile: 100%;
+
+$breakpoint-desktop-max: 1409px;
+$breakpoint-desktop-min: 1410px;
+
+$breakpoint-laptop-max: 1139px;
+$breakpoint-laptop-min: 1140px;
+
+$breakpoint-mobile-max: 959px;
+$breakpoint-mobile-min: 960px;
+
+:root {
+  --risk-disclosure-background: #ffffff;
+  --risk-disclosure-color: #444e5a;
+  --risk-disclosure-button: #00c0ab;
+}
+
+[data-theme='dark'] {
+  --risk-disclosure-background: #444e5a;
+  --risk-disclosure-color: #ffffff;
+  --risk-disclosure-button: #29323d;
+}
