--- packages/fulcrum/src/styles/components/risk-disclosure.scss
@@ -1,109 +1,0 @@
-.disclosure__wrapper {
-  position: relative;
-  background-color: #444e5a;
-  color: #ffffff;
-  box-sizing: border-box;
-
-  @media screen and (min-width: 690px) {
-    width: 650px;
-    padding: 30px 40px;
-    margin: 90px 0;
-  }
-  @media screen and (max-width: 689px) {
-    width: calc(100vw - 60px);
-    padding: 15px 20px;
-    margin: 30px 0;
-  }
-  h2 {
-    line-height: 1.4;
-    margin-bottom: 30px;
-    @media screen and (min-width: 690px) {
-      font-size: 32px;
-      text-align: center;
-    }
-    @media screen and (max-width: 689px) {
-      font-size: 24px;
-    }
-  }
-  h3 {
-    line-height: 1.5;
-    margin-top: 15px;
-    margin-bottom: 10px;
-    @media screen and (min-width: 690px) {
-      font-size: 22px;
-    }
-    @media screen and (max-width: 689px) {
-      font-size: 18px;
-    }
-  }
-  p {
-    line-height: 1.8;
-    margin-bottom: 15px;
-    @media screen and (min-width: 690px) {
-      font-size: 14px;
-    }
-    @media screen and (max-width: 689px) {
-      font-size: 12px;
-    }
-  }
-  a {
-    color: #5b88ff;
-    &:hover {
-      opacity: 0.9;
-    }
-  }
-  ul {
-    list-style: none;
-    padding-left: 15px;
-    li {
-      position: relative;
-      line-height: 2;
-      @media screen and (min-width: 690px) {
-        font-size: 14px;
-      }
-      @media screen and (max-width: 689px) {
-        font-size: 12px;
-      }
-      &:before {
-        content: '';
-        width: 4px;
-        height: 4px;
-        border-radius: 50%;
-        background-color: #56a9ff;
-        position: absolute;
-        left: -15px;
-        @media screen and (min-width: 690px) {
-          top: 14px;
-        }
-        @media screen and (max-width: 689px) {
-          top: 12px;
-        }
-      }
-    }
-  }
-  .disclosure__footer {
-    text-align: right;
-    button {
-      font-size: 15px;
-      line-height: 1.5;
-      color: #ffffff;
-      border-radius: 40px;
-      border: none;
-      outline: none;
-      background-color: #29323d;
-      cursor: pointer;
-      @media screen and (min-width: 690px) {
-        padding: 12px 25px;
-      }
-      @media screen and (max-width: 689px) {
-        padding: 7px 18px;
-      }
-    }
-  }
-  .disclosure__close {
-    cursor: pointer;
-    position: absolute;
-    right: 25px;
-    top: 25px;
-  }
-}
