+++ packages/fulcrum/src/styles/index.scss
@@ -1,3 +1,4 @@
+@import './../../../bzx-common/src/styles/index.scss';
 @import 'abstracts/functions';
 @import 'abstracts/mixins';
 @import 'abstracts/variables';
@@ -17,7 +18,6 @@
 @import 'components/on-chain-indicator';
 @import 'components/preloader';
 @import 'components/provider-menu';
-@import 'components/risk-disclosure';
 @import 'components/rollover-note.scss';
 @import 'components/slider.scss';
 @import 'components/switch-button';
