--- packages/torque/src/styles/components/risk-disclosure.scss
@@ -1,80 +1,0 @@
-.disclosure {
-  position: relative;
-  background-color: #ffffff;
-  color: #444e5a;
-  box-sizing: border-box;
-
-  @media screen and (min-width: 680px) {
-    width: 650px;
-    margin: 90px auto;
-  }
-  @media screen and (max-width: 679px) {
-    width: calc(100vw - 30px);
-    margin: 30px 0;
-  }
-  h3 {
-    font-weight: 800;
-    line-height: 1.5;
-    margin: 15px auto 10px auto;
-    @media screen and (min-width: 690px) {
-      font-size: 18px;
-    }
-    @media screen and (max-width: 689px) {
-      font-size: 16px;
-    }
-  }
-  p {
-    line-height: 1.8;
-    &:not(:last-child) {
-      margin-bottom: 15px;
-    }
-    @media screen and (min-width: 690px) {
-      font-size: 14px;
-    }
-    @media screen and (max-width: 689px) {
-      font-size: 12px;
-    }
-  }
-  a {
-    color: #276bfb;
-    &:hover {
-      opacity: 0.9;
-    }
-  }
-  ul {
-    list-style: none;
-    padding-left: 15px;
-    li {
-      position: relative;
-      line-height: 2;
-      @media screen and (min-width: 690px) {
-        font-size: 14px;
-      }
-      @media screen and (max-width: 689px) {
-        font-size: 12px;
-      }
-      &:before {
-        content: '';
-        width: 4px;
-        height: 4px;
-        border-radius: 50%;
-        background-color: #276bfb;
-        position: absolute;
-        left: -15px;
-        @media screen and (min-width: 690px) {
-          top: 12px;
-        }
-        @media screen and (max-width: 689px) {
-          top: 10px;
-        }
-      }
-    }
-  }
-  .dialog-content {
-    margin-top: 30px;
-  }
-  .dialog-actions {
-    display: flex;
-    justify-content: flex-end;
-  }
-}
