+++ packages/torque/src/styles/index.scss
@@ -1,3 +1,4 @@
+@import './../../../bzx-common/src/styles/index.scss';
 @import 'abstracts/functions';
 @import 'abstracts/mixins';
 @import 'abstracts/variables';
@@ -28,7 +29,6 @@
 @import 'components/ops-selector';
 @import 'components/provider-menu';
 @import 'components/repay-loan-form';
-@import 'components/risk-disclosure';
 @import 'components/loader';
 @import 'components/input-amount';
 @import 'components/progress-fragment';
